Versionado 2 de Microservicios completado y funcional

+++ cita/Cita/src/main/java/com/microservicios_taller/Cita/Cita/repository/CitaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CitaRepository extends JpaRepository<Cita, Long> {
+public interface CitaRepository extends JpaRepository<Cita, Integer> {
 }
