+++ disponibilidad/Disponibilidad/src/main/java/com/microservicios_taller/Disponibilidad/Disponibilidad/repository/DisponibilidadRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DisponibilidadRepository extends JpaRepository<Disponibilidad, Long> {
+public interface DisponibilidadRepository extends JpaRepository<Disponibilidad, Integer> {
 }
